Improve lists and other data structures
Provide more standard-compliant member_to_inst implementation that uses
offsetof. Avoid potential undefined behavior in list_foreach and
list_foreach_rev by avoiding assinging an unaligned pointer value. Use
size_t instead of unsigned long for list length.

+++ uspace/lib/c/generic/adt/list.c
@@ -40,6 +40,7 @@
 
 #include <adt/list.h>
 #include <stdbool.h>
+#include <stdint.h>
 
 /** Check for membership
  *
@@ -99,9 +100,9 @@ void list_splice(list_t *list, link_t *pos)
  * @param list		List to count.
  * @return		Number of items in the list.
  */
-unsigned long list_count(const list_t *list)
+size_t list_count(const list_t *list)
 {
-	unsigned long count = 0;
+	size_t count = 0;
 
 	link_t *link = list_first(list);
 	while (link != NULL) {
